requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
PyJWT>=2.8.0
bcrypt>=4.0.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
aiosqlite>=0.20.0
jinja2>=3.1.4
pyyaml>=6.0
